Solve problem set
Reverted previous changes as header, text, and buttons need to be
classed together. Classing separately has them each treated as their own
flex container and can't be arranged.

Since "content" is now it's own container "close-button" is moved to the
bottom so that is appears after "content" and places it on the right
side.

"Continue" and "cancel" classed together so that they aren't placed in
a column on top of each other.

Flex-basis 0 ensures "content" doesn't take full width of window and
push "cancel" onto next line.